moved dispatch table definition to netcdf_dispatch.h
